brain-link-tracker/backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
attrs==25.3.0
blinker==1.9.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
Flask==3.1.1
flask-cors==6.0.0
Flask-SQLAlchemy==3.1.1
frozenlist==1.7.0
geoip2==5.1.0
greenlet==3.2.4
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
maxminddb==2.8.2
multidict==6.6.4
propcache==0.3.2
psycopg2-binary==2.9.10
python-dotenv==1.1.1
requests==2.32.5
SQLAlchemy==2.0.41
typing_extensions==4.14.0
ua-parser==1.0.1
ua-parser-builtins==0.18.0.post1
urllib3==2.5.0
user-agents==2.2.0
Werkzeug==3.1.3
yarl==1.20.1
